Availability Factor Table

E4ST.setup_table! — Method

setup_table!(config, data, ::Val{:af_table})

Populates the af column of the gen_table.

Updates the generator table with the availability factors provided. By default assigns an availability factor of 1.0 for every generator. See summarize_table(::Val{:af_table}).

Often, generators are unable to generate energy at their nameplate capacity over the course of any given representative hour. This could depend on any number of things, such as how windy it is during a given representative hour, the time of year, the age of the generating unit, etc. The ratio of available generation capacity to nameplate generation capacity is referred to as the availability factor (AF).

The availability factor table includes availability factors for groups of generators specified by any combination of area, genfuel, gentype, year, and hour.

\[P_{G_{g,h,y}} \leq f_{\text{avail}_{g,h,y}} \cdot P_{C{g,y}} \qquad \forall \{g \in \text{generators}, h \in \text{hours}, y \in \text{years} \}\]

E4ST.summarize_table — Method

summarize_table(::Val{:af_table})

column_name	data_type	unit	required	description

area	AbstractString	E4ST.NA	true	The area with which to filter by. I.e. "state". Leave blank to not filter by area.

subarea	AbstractString	E4ST.NA	true	The subarea to include in the filter. I.e. "maryland". Leave blank to not filter by area.

genfuel	AbstractString	E4ST.NA	true	The fuel type that the generator uses. Leave blank to not filter by genfuel.

gentype	String	E4ST.NA	true	The generation technology type that the generator uses. Leave blank to not filter by gentype.

year	E4ST.YearString	E4ST.Year	false	The year to apply the AF's to, expressed as a year string prepended with a "y". I.e. "y2022"

status	Bool	E4ST.NA	false	Whether or not to use this AF adjustment

h_	Float64	E4ST.MWhGeneratedPerMWhCapacity	true	Availability factor of hour _. Include 1 column for each hour in the hours table. I.e. :h1, :h2, ... :hn