Bus Table

E4ST.setup_table! — Method

setup_table!(config, data, ::Val{:bus})

Sets up the bus table.

Makes a :bus_idx to track row numbers.

E4ST.summarize_table — Method

summarize_table(::Val{:bus})

column_name	data_type	unit	required	description

ref_bus	Bool	E4ST.NA	true	Whether or not the bus is a reference bus. There should be a single reference bus for each island.

reg_factor	Float64	E4ST.NA	true	The percentage of generation that dispatches to a cost-of-service regulated market

Additional Columns:

Depending on what you are doing with the model, you may require additional fields associated with buses. These could be things like:

latitude and longitude

The area that a bus is in such as state

The grid that the bus dispatches to