Generator Table

Table representing all existing generators to be modeled.

Note that the generator table does not contain Direct Air Capture, or Storage facilities.

E4ST.summarize_table — Method

summarize_table(::Val{:gen})

column_name	data_type	unit	required	description

bus_idx	Int64	E4ST.NA	true	The index of the bus table that the generator corresponds to

status	Bool	E4ST.NA	false	Whether or not the generator is in service

build_status	InlineStrings.String15	E4ST.NA	true	Whether the generator is built, 'new, or unbuilt. All generators marked new when the gen file is read in will be changed to built. Can also be changed to retired_exog or retired_endog after the simulation is run. See update_build_status!

build_type	AbstractString	E4ST.NA	true	Whether the generator is 'real', 'exog' (exogenously built), or 'endog' (endogenously built)

build_id	AbstractString	E4ST.NA	true	Identifier of the build row. For pre-existing generators not specified in the build file, this is usually left empty

year_on	E4ST.YearString	E4ST.Year	true	The first year of operation for the generator. (For new gens this is also the year it was built)

year_unbuilt	E4ST.YearString	E4ST.Year	false	The latest year the generator was known not to be built. Defaults to year_on - 1. Used for past capex accounting.

econ_life	Float64	E4ST.NumYears	true	The number of years in the economic lifetime of the generator.

year_off	E4ST.YearString	E4ST.Year	true	The first year that the generator is no longer operating in the simulation, computed from the simulation. Leave as y9999 if an existing generator that has not been retired in the simulation yet.

year_shutdown	E4ST.YearString	E4ST.Year	true	The forced (exogenous) shutdown year for the generator. Often equal to the yearon plus the econlife

genfuel	AbstractString	E4ST.NA	true	The fuel type that the generator uses

gentype	String	E4ST.NA	true	The generation technology type that the generator uses

pcap_inv	Float64	E4ST.MWCapacity	true	Original invested nameplate power generation capacity for the generator. This is the original invested capacity of exogenously built generators (even if there have been retirements ), and the original invested capacity in year_on for endogenously built generators.

pcap0	Float64	E4ST.MWCapacity	true	Nameplate power generation capacity for the generator at the start of the simulation

pcap_min	Float64	E4ST.MWCapacity	true	Minimum nameplate power generation capacity of the generator (normally set to zero to allow for retirement)

pcap_max	Float64	E4ST.MWCapacity	true	Maximum nameplate power generation capacity of the generator

vom	Float64	E4ST.DollarsPerMWhGenerated	true	Variable operation and maintenance cost per MWh of generation

fuel_price	Float64	E4ST.DollarsPerMMBtu	false	Fuel cost per MMBtu of fuel used. heat_rate column also necessary when supplying fuel_price

heat_rate	Float64	E4ST.MMBtuPerMWhGenerated	false	Heat rate, or MMBtu of fuel consumed per MWh electricity generated (0 for generators that don't use combustion)

fom	Float64	E4ST.DollarsPerMWCapacityPerHour	true	Hourly fixed operation and maintenance cost for a MW of generation capacity

capex	Float64	E4ST.DollarsPerMWBuiltCapacityPerHour	true	Hourly capital expenditures for a MW of generation capacity. For already-built generators, this is not accounted for in the optimization or accounting. For accounting for investment costs and subsidies in built generators, use past_invest_cost and past_invest_subsidy

transmission_capex	Float64	E4ST.DollarsPerMWBuiltCapacityPerHour	true	Hourly capital expenditures for the transmission supporting a MW of generation capacity

routine_capex	Float64	E4ST.DollarsPerMWCapacityPerHour	true	Routine capital expenditures for a MW of discharge capacity

past_invest_cost	Float64	E4ST.DollarsPerMWCapacityPerHour	false	Investment costs per MW of initial capacity per hour, for past investments

past_invest_subsidy	Float64	E4ST.DollarsPerMWCapacityPerHour	false	Investment subsidies from govt. per MW of initial capacity per hour, for past investments

cf_min	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The minimum capacity factor, or operable ratio of power generation to capacity for the generator to operate. Take care to ensure this is not above the hourly availability factor in any of the hours, or else the model may be infeasible. Set to zero by default.

cf_max	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The maximum capacity factor, or operable ratio of power generation to capacity for the generator to operate

cf_hist	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The historical capacity factor for the generator, or the gentype if no previous data is available. Primarily used to calculate estimate policy value (PTC and EmissionPrice capex_adj)

af	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The availability factor, or maximum available ratio of pewer generation to nameplate capacity for the generator.

emis_co2	Float64	E4ST.ShortTonsPerMWhGenerated	false	The emission rate per MWh of CO2

capt_co2_percent	Float64	E4ST.NA	false	The percentage of co2 emissions captured, to be sequestered.

heat_rate	Float64	E4ST.MMBtuPerMWhGenerated	false	Heat rate, or MMBtu of fuel consumed per MWh electricity generated (0 for generators that don't use combustion)

chp_co2_multi	Float64	E4ST.NA	false	The percentage of CO2 emissions from CHP attributed to the power generation. Used to calculate CO2e

reg_factor	Float64	E4ST.NA	true	The percentage of generation that dispatches to a cost-of-service regulated market

Buildable Generator Specifications

E4ST.summarize_table — Method

summarize_table(::Val{:build_gen})

column_name	data_type	unit	required	description

area	AbstractString	E4ST.NA	true	The area with which to filter by. I.e. "state". Leave blank to not filter by area.

subarea	AbstractString	E4ST.NA	true	The subarea to include in the filter. I.e. "maryland". Leave blank to not filter by area.

build_status	InlineStrings.String15	E4ST.NA	true	Whether the generator is built, 'new, or unbuilt. All generators marked new when the gen file is read in will be changed to built. Can also be changed to retired_exog or retired_endog after the simulation is run. See update_build_status!

build_type	AbstractString	E4ST.NA	true	Whether the generator is 'real', 'exog' (exogenously built), or 'endog' (endogenously built). Should either be exog or endog for buil_gen.

build_id	AbstractString	E4ST.NA	true	Identifier of the build row. Each generator made using this build spec will inherit this build_id

genfuel	AbstractString	E4ST.NA	true	The fuel type that the generator uses. Leave blank to not filter by genfuel.

gentype	String	E4ST.NA	true	The generation technology type that the generator uses. Leave blank to not filter by gentype.

status	Bool	E4ST.NA	false	Whether or not to use this set of characteristics/specs

pcap0	Float64	E4ST.MWCapacity	true	Starting nameplate power generation capacity for the generator. Should be 0 for endog new gens.

pcap_min	Float64	E4ST.MWCapacity	true	Minimum nameplate power generation capacity of the generator (normally set to zero to allow for retirement)

pcap_max	Float64	E4ST.MWCapacity	true	Maximum nameplate power generation capacity of the generator

vom	Float64	E4ST.DollarsPerMWhGenerated	true	Variable operation and maintenance cost per MWh of generation

fuel_price	Float64	E4ST.DollarsPerMMBtu	false	Fuel cost per MMBtu of fuel used. heat_rate column also necessary when supplying fuel_price

fom	Float64	E4ST.DollarsPerMWCapacityPerHour	true	Hourly fixed operation and maintenance cost for a MW of generation capacity

capex	Float64	E4ST.DollarsPerMWBuiltCapacityPerHour	true	Hourly capital expenditures for a MW of generation capacity

transmission_capex	Float64	E4ST.DollarsPerMWBuiltCapacityPerHour	true	Hourly capital expenditures for the transmission supporting a MW of generation capacity

routine_capex	Float64	E4ST.DollarsPerMWCapacityPerHour	true	Routing capital expenditures for a MW of discharge capacity

cf_min	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The minimum capacity factor, or operable ratio of power generation to capacity for the generator to operate. Take care to ensure this is not above the hourly availability factor in any of the hours, or else the model may be infeasible. Set to zero by default.

cf_max	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The maximum capacity factor, or operable ratio of power generation to capacity for the generator to operate

cf_hist	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The historical capacity factor for the generator or the gentype. Primarily used to calculate estimate policy value (PTC and EmissionPrice capex_adj)

year_on	E4ST.YearString	E4ST.Year	true	The first year of operation for the generator. (For new gens this is also the year it was built). Endogenous unbuilt generators will be left blank

econ_life	Float64	E4ST.NumYears	true	The number of years in the economic lifetime of the generator.

age_shutdown	Float64	E4ST.NumYears	true	The age at which the generator is no longer operating. I.e. if year_on = y2030 and age_shutdown = 20, then capacity will be 0 in y2040.

year_on_min	E4ST.YearString	E4ST.Year	true	The first year in which a generator can be built/come online (inclusive). Generators with no restriction and exogenously built gens will be left blank

year_on_max	E4ST.YearString	E4ST.Year	true	The last year in which a generator can be built/come online (inclusive). Generators with no restriction and exogenously built gens will be left blank

emis_co2	Float64	E4ST.ShortTonsPerMWhGenerated	false	The CO2 emission rate of the generator, in short tons per MWh generated. This is the net emissions. (i.e. not including captured CO2 that gets captured)

capt_co2_percent	Float64	E4ST.NA	false	The percentage of co2 emissions captured, to be sequestered.