Hours Table

E4ST.summarize_table — Method

summarize_table(::Val{:hours})

column_name	data_type	unit	required	description

hours	Float64	E4ST.Hours	true	The number of hours spent in each representative hour over the course of a year (must sum to 8760)

E4ST.get_hour_weights — Function

get_hour_weights(data) -> weights

get_hour_weights(data, hour_idxs) -> weights (view)

Returns the number of hours in a year spent at each representative hour

E4ST.get_hour_weight — Function

get_hour_weight(data, hour_idx)

Returns the number of hours in a year spent at the hour_idx representative hour