Container

E4ST.Container — Type

abstract type Container

Abstract type for containers that can be indexed by year and time. i.e. c[yr_idx, hr_idx] will work, even if c contains a single number, a number for each year, a number for each hour, or a number for each year and hour.

Missing docstring.

Missing docstring for ByNothing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ByYear. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ByHour. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ByYearAndHour. Check Documenter's build log for details.

E4ST.to_container — Function

to_container(v) -> cv::Vector{Container}

Converts v to a Vector{Container}

E4ST.to_container! — Function

to_container!(table, col_name)

Converts table[!, :col_name] to a Vector{Container}.