Yearly Adjustments

E4ST.AdjustYearly — Type

AdjustYearly(;file, name)

Adjusts tables and parameters by year. Stores the table stored in file into data[name].

column_name	data_type	unit	required	description

table_name	AbstractString	E4ST.NA	true	The name of the table to adjust. Leave blank if this adjustment is intended for a variable in data

variable_name	AbstractString	E4ST.NA	true	The name of the variable/column to adjust

operation	E4ST.Operation	E4ST.NA	true	The operation to perform. Could be add, scale, or set.

filter_	String	E4ST.NA	true	There can be multiple filter conditions - filter1, filter2, etc. It denotes a comparison used for selecting the table rows to apply the adjustment to. See parse_comparison for examples

year_col	String	E4ST.NA	false	Optional, the column for which the adjustment applies. For example, use year_on to set a cost for the lifetime of a generator. Leave blank (default) for to apply the adjustment by the simulation year value.

status	Bool	E4ST.NA	false	Whether or not to use this adjustment

y_	Float64	E4ST.Ratio	true	Value to adjust by for each year. Include a column for each year in the hours table. I.e. :y2020, :y2030, etc

Hourly Adjustments

E4ST.AdjustHourly — Type

AdjustHourly(;file, name)

Adjusts tables and parameters by hour. Stores the table stored in file into data[name].

column_name	data_type	unit	required	description

table_name	AbstractString	E4ST.NA	true	The name of the table to adjust. Leave blank if this adjustment is intended for a variable in data

variable_name	AbstractString	E4ST.NA	true	The name of the variable/column to adjust

operation	E4ST.Operation	E4ST.NA	true	The operation to perform. Could be add, scale, or set.

filter_	String	E4ST.NA	true	There can be multiple filter conditions - filter1, filter2, etc. It denotes a comparison used for selecting the table rows to apply the adjustment to. See parse_comparison for examples

year	String	E4ST.Year	true	The year to adjust, expressed as a year string prepended with a "y". I.e. "y2022". Leave blank to adjust all years

status	Bool	E4ST.NA	false	Whether or not to use this adjustment

h_	Float64	E4ST.Ratio	true	Value to adjust by for each hour. Include a column for each hour in the hours table. I.e. :h1, :h2, ... :hn

Adjustments By Age

E4ST.AdjustByAge — Type

AdjustByAge(;file, name)

Adjusts tables and parameters by year. Stores the table stored in file into data[name].

column_name	data_type	unit	required	description

table_name	AbstractString	E4ST.NA	true	The name of the table to adjust. Leave blank if this adjustment is intended for a variable in data

variable_name	AbstractString	E4ST.NA	true	The name of the variable/column to adjust

operation	E4ST.Operation	E4ST.NA	true	The operation to perform. Could be add, scale, or set.

filter_	String	E4ST.NA	true	There can be multiple filter conditions - filter1, filter2, etc. It denotes a comparison used for selecting the table rows to apply the adjustment to. See parse_comparison for examples

status	Bool	E4ST.NA	false	Whether or not to use this adjustment

age_type	String	E4ST.NA	true	The type of age specified, can be exact, after, or trigger. If exact, then adjustment is applied only when the age in question is between [age, age+1). If trigger, then adjustment is applied for the first simulation year for which the age has been exceeded. If after, then adjustment is applied on [age, Inf)

age	Float64	E4ST.NumYears	true	The age at which to apply this adjustment. Applies depending on age_type

value	Float64	E4ST.NA	true	Value to adjust by.