Annual Capacity Factor Limit

E4ST.AnnualCapacityFactorLimit — Type

struct AnnualCapacityFactorLimit <: Modification

AnnualCapacityFactorLimit(;name, file)

Sets annual capacity factor limits for generators. Annual capacity factor is defined as the total energy generated in a year divided by the total amount of energy capacity (power capacity times the number of hours in a year).

modify_raw_data! - Loads in a table from file, stores it into data[<name>]. See summarizetable(::Val{:annualcf_lim})

modify_model! - sets up the following constraints and expressions

Sets up expression model[:egen_gen_annual] (ngen x nhr) for annual energy generation for each generator

Creates constraint model[:cons_<name>_min] for each generator covered by each row of the table specified in file, if the annual_cf_min column is given.

Creates constraint model[:cons_<name>_max] for each generator covered by each row of the table specified in file, if the annual_cf_max column is given.

E4ST.summarize_table — Method

summarize_table(::Val{:annual_cf_lim})

column_name	data_type	unit	required	description

genfuel	AbstractString	E4ST.NA	false	The fuel type that the generator uses. Leave blank to not filter by genfuel.

gentype	String	E4ST.NA	false	The generation technology type that the generator uses. Leave blank to not filter by gentype.

area	AbstractString	E4ST.NA	false	The area with which to filter by. I.e. "state". Leave blank to not filter by area.

subarea	AbstractString	E4ST.NA	false	The subarea to include in the filter. I.e. "maryland". Leave blank to not filter by area.

filter_	String	E4ST.NA	false	There can be multiple filter conditions - filter1, filter2, etc. It denotes a comparison used for selecting the table rows to apply the adjustment to. See parse_comparison for examples

status	Bool	E4ST.NA	false	Whether or not to use this limit

annual_cf_min	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The minimum annual capacity factor ∈ (0,1]. If outside these bounds, not set. Be very careful - easy to make model infeasible if contradictory to availability factors.

annual_cf_max	Float64	E4ST.MWhGeneratedPerMWhCapacity	false	The maximum annual capacity factor ∈ [0,1). If outside these bounds, not set.