Branch Table

Table representing all existing branches (AC transmission lines) to be modeled. See also DCLine

Note that the Branch Table does not contain DC transmission lines.

E4ST.setup_table! — Method

setup_table!(config, data, ::Val{:branch})

Sets up the branch table.

Flips f_bus_idx and t_bus_idx so that f_bus_idx < t_bus_idx

Makes bus[:connectedbranchidxs] which contains a vector of the signed index of each branch leaving that bus. (+ for f_bus_idx, - for to_bus_idx).

E4ST.summarize_table — Method

summarize_table(::Val{:branch})

column_name	data_type	unit	required	description

f_bus_idx	Int64	E4ST.NA	true	The index of the bus table that the branch originates from

t_bus_idx	Int64	E4ST.NA	true	The index of the bus table that the branch goes to

status	Bool	E4ST.NA	false	Whether or not the branch is in service

x	Float64	E4ST.PU	true	Per-unit reactance of the line (resistance assumed to be 0 for DC-OPF)

pflow_max	Float64	E4ST.MWFlow	true	Maximum power flowing through the branch