Logging

In general, E4ST piggy-backs off of the Logging.jl interface. To make a log statement, simply write:

@info "Something informative"
@warn "Some sort of warning"
@debug "Debugging information, not printed unless config[:logging] == \"debug\""

Documentation

E4ST.start_logging! — Method

start_logging!(config)

Starts logging according to config[:logging]. Possible options for config[:logging]:

true (default): logs @info, @warning, and @error messages to config[:out_path]/E4ST.log

"debug" - logs @debug, @info, @warning, and @error messages to config[:out_path]/E4ST.log

false - no logging

To log things, you can use @info, @warn, or @debug as defined in Logging.jl. Or you can use a convenience method for logging a header, log_header

To stop the logger and close its io stream, see stop_logging!(config)

E4ST.stop_logging! — Method

stop_logging!(config)

Stops logging to console, closes the io stream of the current logger.

E4ST.log_header — Method

log_header(header)

Logs a 3-line header string by calling @info header_string(header)

E4ST.log_start — Function

log_start(config)

Logs any necessary info at the beginning of a run of E4ST

E4ST.header_string — Function

header_string(header) -> s

Returns a 3-line header string

E4ST.time_string — Function

time_string() -> s

Returns a time string in the format "yymmdd_HHMMSS"

E4ST.date_string — Function

date_string() -> s

Returns a date string in the format "yymmdd"