Overview

E4ST.run_e4st — Function

run_e4st(config) -> out_path

run_e4st(filename(s)) -> out_path

Top-level function for running E4ST. Here is a general overview of what happens:

Book-keeping

read_config(config_file) - loads in the config from file, if not passed in directly.

save_config(config) - the config is saved to config[:out_path]

start_logging!(config) - Logging is started, some basic information is logged via log_start.

Load Input Data

read_data(config) - The data is loaded in from files specified in the config.

Construct JuMP Model and optimize

setup_model(config, data) - The model (a JuMP Model) is set up.

JuMP.optimize!(model) - The model is optimized.

Process Results

parse_results!(config, data, model) - Retrieves all necessary values and shadow prices from model, storing them into data[:results][:raw], (see get_raw_results and get_results) and saves data if config[:save_data_parsed] is true (default is true). This is mostly stored in case the results processing throws an error before completion. That way, there is no need to re-run the model.

process_results!(config, data) - Calls modify_results!(mod, config, data) for each mod in the config. Saves data if config[:save_data_processeded] is true (default is true)

Iterate, running more simulations as needed.

See Iterable and read_config for more information.