DC Transmission Lines

E4ST.DCLine — Type

struct DCLine <: Modification
    
DCLine(;file)

This Modification takes in a file representing the dc lines to add to the model. See summarize_table(::Val{:dc_line}) for info on the table.

This creates a single variable for each dc line (at each point in time), and adds it to pbal_bus of the t_bus_idx, and subtracts it from the f_bus_idx. Represents a lossless transfer of power, ignoring voltage angle requirements.

Keyword Argument:

file - the filename of the dc line table, where each row represents a different dc line. See also summarize_table(::Val{:dc_line})

Table Added:

dc_line - this table contains information about each dc line, with rows for indexes of the buses power is flowing to/from and the amount of power flow.

Model Modification:

Variables

(:dc_line, :pflow_dc) - the amount of power flowing through a dc line

Interfaces Implemented

modify_raw_data!(mod::DCLine, config, data) - loads mod.file => data[:dc_line]

modify_model!(mod::DCLine, config, data, model) - Add dc lines to the model from data[:dc_lines], creating pflow_dc variables, and adding/subtracting to the corresponding pflow_bus variables.

E4ST.modify_raw_data! — Method

modify_raw_data!(mod::DCLine, config, data)

Loads mod.file => data[:dc_line]

E4ST.modify_model! — Method

modify_model!(mod::DCLine, config, data, model)

Add dc lines to the model from data[:dc_lines], creating pflow_dc variables, and adding/subtracting to the corresponding pflow_bus variables.

E4ST.summarize_table — Method

summarize_table(::Val{:dc_line})

column_name	data_type	unit	required	description

f_bus_idx	Int64	E4ST.NA	true	The index of the bus table that the line originates from

t_bus_idx	Int64	E4ST.NA	true	The index of the bus table that the line goes to

status	Bool	E4ST.NA	false	Whether or not the dc line is in service

pflow_max	Float64	E4ST.MWFlow	true	Maximum power flowing through the dc line in the direction f_bus_idx to t_bus_idx

pflow_max_reverse	Float64	E4ST.MWFlow	false	Maximum power flowing through the dc line in the direction t_bus_idx to f_bus_idx. Defaults to 0 if not provided